+++ src/Router.jsx
@@ -5,9 +5,9 @@ import SearchPage from './pages/main/Seach';
 import FavoritesPage from './pages/favorites/Favorites';
 import SignInPage from './pages/auth/SignIn';
 import SignUpPage from './pages/auth/SignUp';
-import ArtworkUploadPage from './pages/artworkUpload/ArtworkUpload';
-import ArtworkDetailPage from './pages/artworkDetail/ArtworkDetail';
-import MyArtworkDetailPage from './pages/artworkDetail/MyArtworkDetail';
+import ArtworkUploadPage from './pages/artwork/artworkUpload/ArtworkUpload';
+import ArtworkDetailPage from './pages/artwork/artworkDetail/ArtworkDetail';
+import MyArtworkDetailPage from './pages/artwork/artworkDetail/MyArtworkDetail';
 import ArtistPage from './pages/artist/Artist';
 import MyReview from './pages/myPage/MyReview';
 import MyPage from './pages/myPage/MyPage';
