+++ src/pages/artwork/artworkDetail/ArtworkDetail.style.js
